+++ packages/ak-button/test/index.js
@@ -84,9 +84,9 @@ describe('ak-button', () => {
         });
       });
 
-      [APPEARANCE.PRIMARY, APPEARANCE.SUBTLE, APPEARANCE.SELECTED].forEach(appearanceName => {
+      [APPEARANCE.PRIMARY, APPEARANCE.SUBTLE, APPEARANCE.COMPACT].forEach(appearanceName => {
         describe(appearanceName, () => {
-          const selector = `.${classKeys.container} .${classKeys[appearanceName]}`;
+          const selector = `.${classKeys.button}.${classKeys[appearanceName]}`;
           beforeEach(() =>
             props(component, { appearance: appearanceName })
           );
@@ -103,8 +103,24 @@ describe('ak-button', () => {
       });
     });
 
+    describe('selected', () => {
+      const selector = `.${classKeys.button}.${classKeys.selected}`;
+      beforeEach(() =>
+        props(component, { selected: true })
+      );
+
+      it('button should have selected class', () =>
+        expect(shadowDomQuery(component, selector)).not.to.be.null
+      );
+
+      it('button should not have selected class after it is removed', () => {
+        props(component, { selected: false });
+        expect(shadowDomQuery(component, selector)).to.be.null;
+      });
+    });
+
     describe('disabled', () => {
-      const selector = `.${classKeys.container} button[disabled]`;
+      const selector = `.${classKeys.button}[disabled]`;
       beforeEach(() =>
         props(component, { disabled: true })
       );
@@ -114,7 +130,7 @@ describe('ak-button', () => {
       );
 
       it('button should override any other class', () => {
-        props(component, { disabled: true, appearance: 'selected' });
+        props(component, { disabled: true, selected: true });
         const buttonClasses = getShadowButtonElem(component).classList;
         expect(buttonClasses).to.have.lengthOf(2);
         expect(containsClass(buttonClasses, classKeys.button, classKeys.disabled)).to.be.true;